[DEV-3296] moving fetchAwardDownloadFile to class method for testability
